base(code motion): move `calculate_room_name` into the one file where it's used
